[refactor] 로그 인터셉터 수정
로깅 박스 생성
